+++ src/components/Forms/EditDestinationsForm.js
@@ -69,16 +69,21 @@ const EditDestinationsForm = ({ mailoutDetails, mailoutDestinationsEdit, handleB
   const [error, setError] = useState(null);
   const currentListingStatus = mailoutDetails?.listingStatus;
   const isCampaign = mailoutDetails?.subtype === 'campaign';
+  const isGeneralCampaign = mailoutDetails?.publishedTags?.includes('general');
   const isCalculationDeferred = mailoutDetails?.mailoutStatus === 'calculation-deferred';
 
   const [destinationsOptionsMode, setDestinationsOptionsMode] = useState(
-    mailoutDetails.destinationsOptions?.mode || isCampaign ? 'manual' : 'ai'
+    (mailoutDetails.destinationsOptions?.mode || isCampaign) && !isGeneralCampaign
+      ? 'manual'
+      : isGeneralCampaign
+      ? 'userUploaded'
+      : 'ai'
   );
-  const [saveDetails, setSaveDetails] = useState({
-    destinationsOptionsMode:
-      mailoutDetails.destinationsOptions?.mode || isCampaign ? 'manual' : 'ai',
-    ready: false,
-  });
+  const [saveDetails, setSaveDetails] = useState(
+    mailoutDetails.destinationsOptions?.mode || isCampaign
+      ? { destinationsOptionsMode: 'manual', ready: false }
+      : { destinationsOptionsMode: 'ai', ready: true }
+  );
 
   const [numberOfDestinations, setNumberOfDestinations] = useState(mailoutDetails.mailoutSize);
 
@@ -234,7 +239,11 @@ const EditDestinationsForm = ({ mailoutDetails, mailoutDestinationsEdit, handleB
 
   const handleSubmitClick = async () => {
     setSaving(true);
-    if (!saveDetails || !saveDetails.ready) return; // somehow got here, but not ready
+    if (!saveDetails || !saveDetails.ready) {
+      // somehow got here, but not ready
+      setSaving(false);
+      return;
+    }
     try {
       if (saveDetails.destinationsOptionsMode === 'ai') {
         let path = `/api/user/mailout/${mailoutDetails._id}/edit/mailoutSize`;
@@ -272,10 +281,9 @@ const EditDestinationsForm = ({ mailoutDetails, mailoutDestinationsEdit, handleB
         const formData = new FormData();
         formData.append('destinations', csvFile);
         if (!isCsvBrivityFormat) {
-          if (firstNameColumn && firstNameColumn !== null)
+          if (typeof firstNameColumn === 'number')
             formData.append('firstNameColumn', firstNameColumn);
-          if (lastNameColumn && lastNameColumn !== null)
-            formData.append('lastNameColumn', lastNameColumn);
+          if (typeof lastNameColumn === 'number') formData.append('lastNameColumn', lastNameColumn);
           formData.append('deliveryLineColumn', deliveryLineColumn);
           formData.append('cityColumn', cityColumn);
           formData.append('stateColumn', stateColumn);