+++ src/components/Navigation.js
@@ -6,6 +6,7 @@ import React, { createRef, useEffect, useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useDispatch } from 'react-redux';
 import { resetMailout } from '../store/modules/mailout/actions';
+import { faFacebookF } from '@fortawesome/free-brands-svg-icons';
 
 import { StepLayout, StepsLayout, MobileDisabledLayout, NavigationLayout } from '../layouts';
 import { Dimmer, Menu, Initials, Icon, Step } from './Base';
@@ -65,12 +66,20 @@ const menuP = {
 const StyledCog = styled(Cog)`
   width: 15px;
   height: 20px;
-  margin-left: 12px;
+  margin-left: 14px;
   margin-top: 17px;
 `;
 
 const StyledIcon = styled(FontAwesomeIcon)`
   margin: 0em 1em 0em 0.65em;
+`;
+
+const ArchiveStyledIcon = styled(FontAwesomeIcon)`
+  margin: 0em 1.2em 0em 0.65em;
+`;
+
+const FacebookStyledIcon = styled(FontAwesomeIcon)`
+  margin: 0em 1.3em 0em 0.86em;
 `;
 
 export default () => {
@@ -127,6 +136,7 @@ export default () => {
   const mailoutUpdateMailoutSizePendingState = useSelector(
     store => store.mailout.updateMailoutSizePending
   );
+  const adProduct = useSelector(store => store.onLogin.permissions?.adProduct);
 
   const profiles = [];
 
@@ -337,6 +347,32 @@ export default () => {
 
           <Menu.Item
             as={Link}
+            name="listings"
+            active={activeItem === '/listings'}
+            to="/listings"
+            style={menuItemStyles}
+            onClick={mobileCollapse}
+          >
+            <MobileDisabledLayout>
+              <StyledIcon icon="home" className="iconWithStyle" /> Listings
+            </MobileDisabledLayout>
+          </Menu.Item>
+          {adProduct && (
+            <Menu.Item
+              as={Link}
+              name="ads"
+              active={activeItem === '/ads'}
+              to="/ads"
+              style={menuItemStyles}
+              onClick={mobileCollapse}
+            >
+              <MobileDisabledLayout>
+                <FacebookStyledIcon icon={faFacebookF} className="facebookIconWithStyle" /> Paid Ads
+              </MobileDisabledLayout>
+            </Menu.Item>
+          )}
+          <Menu.Item
+            as={Link}
             name="archived"
             active={activeItem === '/dashboard/archived'}
             to="/dashboard/archived"
@@ -344,7 +380,7 @@ export default () => {
             onClick={mobileCollapse}
           >
             <MobileDisabledLayout>
-              <StyledIcon icon="archive" className="iconWithStyle" /> Archive
+              <ArchiveStyledIcon icon="archive" className="archvieIconWithStyle" /> Archive
             </MobileDisabledLayout>
           </Menu.Item>